Move `workflow` sub-package out of Triton directory
